requirements.txt
tensorflow-cpu
flask
numpy
pandas
scikit-learn
matplotlib
scipy
gunicorn
